requirements.txt
asgiref==3.3.1
Django==2.2.10
djangorestframework==3.12.3
pytz==2021.1
sqlparse==0.4.1
